refactor(auv-cli-invoke): own invoke request models
