+++ src/model.rs
@@ -1,69 +1,2 @@
-// File: src/model.rs
-use auv_tracing_driver::trace::{ArtifactRecordV1Alpha1, SpanId};
-use std::collections::BTreeMap;
-use std::path::PathBuf;
-use std::process;
-use std::sync::atomic::{AtomicU64, Ordering};
-
+pub use auv_cli_invoke::{ExecutionTarget, InvokeRequest, InvokeResult, RunStatus};
 pub use auv_tracing_driver::{AuvResult, now_millis};
-
-static RUN_ID_COUNTER: AtomicU64 = AtomicU64::new(0);
-
-#[derive(Clone, Debug, Default)]
-pub struct ExecutionTarget {
-  pub application_id: Option<String>,
-  pub target_label: Option<String>,
-}
-
-#[derive(Clone, Debug, Default)]
-pub struct InvokeRequest {
-  pub command_id: String,
-  pub target: ExecutionTarget,
-  pub inputs: BTreeMap<String, String>,
-  pub dry_run: bool,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum RunStatus {
-  Completed,
-  Failed,
-}
-
-impl RunStatus {
-  pub fn as_str(&self) -> &'static str {
-    match self {
-      Self::Completed => "completed",
-      Self::Failed => "failed",
-    }
-  }
-}
-
-#[derive(Clone, Debug)]
-pub struct InvokeResult {
-  pub run_id: String,
-  pub producer_span_id: SpanId,
-  pub status: RunStatus,
-  pub output_summary: String,
-  pub signals: BTreeMap<String, String>,
-  pub artifacts: Vec<ArtifactRecordV1Alpha1>,
-  pub artifact_paths: Vec<PathBuf>,
-  pub failure_message: Option<String>,
-}
-
-pub fn new_run_id() -> String {
-  let sequence = RUN_ID_COUNTER.fetch_add(1, Ordering::Relaxed);
-  format!("run_{}_{}_{}", now_millis(), process::id(), sequence)
-}
-
-#[cfg(test)]
-mod tests {
-  use super::new_run_id;
-
-  #[test]
-  fn new_run_id_is_unique_within_process() {
-    let first = new_run_id();
-    let second = new_run_id();
-
-    assert_ne!(first, second);
-  }
-}
